requirements.txt
fastapi
uvicorn
smartapi-python
pyotp
requests
logzero
python-multipart
pydantic